requirements.txt
qdrant-client>=1.9.2
requests>=2.32.3
openai>=1.66.3
python-dotenv>=1.0.1
PyMuPDF>=1.23.19
pypdf
langchain_openai>=0.3.9
langchain-qdrant>=0.1.0
langchain_community>=0.0.30